backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.0.0

# QuestDB
questdb>=1.1.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pydantic[email]>=2.5.0

# HTTP
httpx>=0.27.0

# Task Queue
celery>=5.3.0

# Utils
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Dev
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
ruff>=0.3.0
mypy>=1.8.0
